Move mangle/unmangle to own module
